users, user_sessions controllers in new namespace
