+++ app/controllers/application_controller.rb
@@ -10,7 +10,7 @@ class ApplicationController < ActionController::Base
   private
 
   def not_authenticated
-    redirect_to login_path, alert: t('application.not_authenticated.need_login')
+    redirect_to home_login_path, alert: t('application.not_authenticated.need_login')
   end
 
   def set_locale
