+++ app/controllers/dashboard/user_session_controller.rb
@@ -1,0 +1,8 @@
+module Dashboard
+  class UserSessionsController < ApplicationController
+    def destroy
+      logout
+      redirect_to login_path, notice: t('.logout')
+    end
+  end
+end
