+++ app/controllers/user_sessions_controller.rb
@@ -1,19 +1,1 @@
-class UserSessionsController < ApplicationController
-  skip_before_action :require_login
-  def new
-  end
 
-  def create
-    if user = login(params[:user_session][:email], params[:user_session][:password])
-      redirect_back_or_to root_path, notice: t('.success')
-    else
-      flash.now[:alert] = t('.fail')
-      render action: 'new'
-    end
-  end
-
-  def destroy
-    logout
-    redirect_to login_path, notice: t('.logout')
-  end
-end
