+++ app/views/dashboard/users/edit.html.erb
@@ -1,0 +1,2 @@
+<p><%= t('.edit_user') %></p>
+<%= render partial: 'shared/users_form', locals: { namespace: :dashboard } %>
