+++ app/views/home/users/new.html.erb
@@ -1,0 +1,2 @@
+<p><%= t('.sign_up') %></p>
+<%= render partial: 'shared/users_form', locals: { namespace: :home } %>
