+++ app/views/layouts/_header.html.erb
@@ -18,15 +18,15 @@
           <li class="dropdown">
             <%= link_to "#{current_user.email}", '#', class: 'dropdown-toggle', 'data-toggle' => 'dropdown' %>
             <ul class="dropdown-menu">
-              <li><%= link_to t('.user_edit'), edit_user_path(current_user) %></li>
-              <li><%= link_to t('.exit'), :logout, method: :post %></li>
+              <li><%= link_to t('.user_edit'), edit_dashboard_user_path(current_user) %></li>
+              <li><%= link_to t('.exit'), dashboard_logout_path, method: :post %></li>
             </ul>
           </li>
         <% else %>
-          <li><%= link_to t('.lang_en'), login_path(locale: 'en') %></li>
-          <li><%= link_to t('.lang_ru'), login_path(locale: 'ru') %></li>
-          <li><%= link_to t('.sign_up'), :signup %></li>
-          <li><%= link_to t('.login'), :login %></li>
+          <li><%= link_to t('.lang_en'), home_login_path(locale: 'en') %></li>
+          <li><%= link_to t('.lang_ru'), home_login_path(locale: 'ru') %></li>
+          <li><%= link_to t('.sign_up'), home_signup_path %></li>
+          <li><%= link_to t('.login'), home_login_path %></li>
         <% end %>
       </ul>
     </div>
