--- app/views/users/edit.html.erb
@@ -1,2 +1,0 @@
-<p><%= t('.edit_user') %></p>
-<%= render 'form' %>
