--- app/views/users/new.html.erb
@@ -1,2 +1,0 @@
-<p><%= t('.sign_up') %></p>
-<%= render 'form' %>
