+++ config/routes.rb
@@ -1,20 +1,27 @@
 Rails.application.routes.draw do
   scope '(:locale)', locale: /ru|en/ do
+    namespace :home do
+      resources :users, only: [:new, :create]
+      get 'signup' => 'users#new' # , :as => :signup
+      get 'login' => 'user_sessions#new' # , :as => :login
+    end
+
+    namespace :dashboard do
+      resources :users, only: [:edit, :update]
+      post 'logout' => 'user_sessions#destroy'
+    end
+
     get 'oauths/oauth'
 
     get 'oauths/callback'
 
     resources :user_sessions
-    resources :users
     resources :cards
     resources :decks
 
     root to: 'flashcards#index'
     post 'compare' => 'flashcards#compare'
 
-    get 'login' => 'user_sessions#new', :as => :login
-    post 'logout' => 'user_sessions#destroy', :as => :logout
-    get 'signup' => 'users#new', :as => :signup
 
     post 'oauth/callback' => 'oauths#callback'
     get 'oauth/callback' => 'oauths#callback'
